refactor(toggle): split to Checkbox, Radio and Switch components
