+++ boilerplate/app/components/Toggle/Checkbox.tsx
@@ -1,0 +1,110 @@
+import React from "react"
+import { Image, ImageStyle, StyleProp, View, ViewStyle } from "react-native"
+import Animated, { useAnimatedStyle, withTiming } from "react-native-reanimated"
+import { colors } from "../../theme"
+import { iconRegistry, IconTypes } from "../Icon"
+import { $inputOuterBase, BaseToggleInputProps, ToggleProps, Toggle } from "./Toggle"
+
+export interface CheckboxToggleProps extends ToggleProps<CheckboxInputProps> {
+  /**
+   * Optional style prop that affects the Image component.
+   */
+  inputDetailStyle?: ImageStyle
+  /**
+   * Checkbox-only prop that changes the icon used for the "on" state.
+   */
+  checkboxIcon?: IconTypes
+}
+
+interface CheckboxInputProps extends BaseToggleInputProps<CheckboxToggleProps> {
+  checkboxIcon?: CheckboxToggleProps["checkboxIcon"]
+}
+
+export function Checkbox(props: CheckboxToggleProps) {
+  return <Toggle {...props} ToggleInput={CheckboxInput} />
+}
+
+/**
+ * @param {ToggleInputProps} props - The props for the `Checkbox` component.
+ * @returns {JSX.Element} The rendered `Checkbox` component.
+ */
+function CheckboxInput(props: CheckboxInputProps) {
+  const {
+    on,
+    status,
+    disabled,
+    checkboxIcon,
+    outerStyle: $outerStyleOverride,
+    innerStyle: $innerStyleOverride,
+    detailStyle: $detailStyleOverride,
+  } = props
+
+  const offBackgroundColor = [
+    disabled && colors.palette.neutral400,
+    status === "error" && colors.errorBackground,
+    colors.palette.neutral200,
+  ].filter(Boolean)[0]
+
+  const outerBorderColor = [
+    disabled && colors.palette.neutral400,
+    status === "error" && colors.error,
+    !on && colors.palette.neutral800,
+    colors.palette.secondary500,
+  ].filter(Boolean)[0]
+
+  const onBackgroundColor = [
+    disabled && colors.transparent,
+    status === "error" && colors.errorBackground,
+    colors.palette.secondary500,
+  ].filter(Boolean)[0]
+
+  const iconTintColor = [
+    disabled && colors.palette.neutral600,
+    status === "error" && colors.error,
+    colors.palette.accent100,
+  ].filter(Boolean)[0]
+
+  return (
+    <View
+      style={[
+        $inputOuter,
+        { backgroundColor: offBackgroundColor, borderColor: outerBorderColor },
+        $outerStyleOverride,
+      ]}
+    >
+      <Animated.View
+        style={[
+          $checkboxInner,
+          { backgroundColor: onBackgroundColor },
+          $innerStyleOverride,
+          useAnimatedStyle(() => ({ opacity: withTiming(on ? 1 : 0) }), [on]),
+        ]}
+      >
+        <Image
+          source={checkboxIcon ? iconRegistry[checkboxIcon] : iconRegistry.check}
+          style={[
+            $checkboxDetail,
+            !!iconTintColor && { tintColor: iconTintColor },
+            $detailStyleOverride,
+          ]}
+        />
+      </Animated.View>
+    </View>
+  )
+}
+
+const $checkboxInner: ViewStyle = {
+  width: "100%",
+  height: "100%",
+  alignItems: "center",
+  justifyContent: "center",
+  overflow: "hidden",
+}
+
+const $checkboxDetail: ImageStyle = {
+  width: 20,
+  height: 20,
+  resizeMode: "contain",
+}
+
+const $inputOuter: StyleProp<ViewStyle> = [$inputOuterBase, { borderRadius: 4 }]
